requirements.txt
aws-cdk-lib==2.110.1
constructs>=10.0.0,<11.0.0
cdk-nag==2.27.200
awscli-local==0.21.1
boto3==1.29.6
botocore==1.32.6